convert the generated HyphenTable.ml to revised syntax
OMake tries to scan the dependencies of HyphenTable.ml using a camlp4
preprocessor that assumes revised syntax. I don't know how to fix
that, so I just converted the code generator to revised syntax.

This means that I had to remove another part of the OMakefile that
would actually build the file using standard syntax. Now it's using
the revised preprocessor as the rest of the codebase.