include a minimal build.sh script:
    mkdir lib
    omake Runtime/kpathsea-stubs.o Runtime/freetype-stubs.o Unicode/Tables.o
    omake

I think that lib/ was lost in the git import, as it is an empty directory.

I have no idea why those .o files have to be built manually. Running
'omake' without building them first fails on them, and this is how
I collected the list of files that need to be built manually. They all
correspond to files that I had to hack to get to build.